refactor: split deps command modules
