+++ crates/fabrik-cli/src/commands/deps/graph.rs
